Fix for macro execution problem.
Macros weren't executing properly because swapping to and from normal mode was being
interpereted as separate keystrokes. The root cause was a fix I'd put in to make vim
treat alt as alt in iTerm. The reason for this fix was so that I could navigate between
spilts with <M-[hljk]>. Given that's my only alt shortcut, the easiest solution was to
change it to <C-[hljk]> and drop the alt remapper