Update to allow script to work under sudo.
So, $HOME/similar are set to the original user when the script is called using sudo, which it is under chef. The only way around this I could find was to get the effective UID and then use the Ruby 'etc' module to parse /etc/passwd and get the username from it and use ~username instead of $HOME.

Computers.